Use Apache Commons VFS2 to support multiple inputs
Previously, the old and new inputs to delta-diff had to
be directories.  This change allows for either new or old
to be a zip file.  It's hardcoded to assume zip when the
old or new directory is not a directory.

This only works for delta-diff right now, but it may be
possible to make delta-patch do something similar.

This is relatively untested, the provided test passes and
tests with a 'old' zip to 'new' directory seem to produce
the expected delta.